Handle two elliptical arc implementation note details
As mentioned in issue #13 `grammar.peg` is a completely literal
implementation of the SVG 1.1 grammar found in
[8.3.9 The grammar for path
data](https://www.w3.org/TR/SVG/paths.html#PathDataBNF).
However the SVG spec's grammar does not reflect the appendix F.5 and F.6
implementation notes for path strings, so a strictly derived grammar
does not match the complete SVG implementation requirements.

This change covers two of the details mentioned in appendix
[F.6 Elliptical arc implementation
notes](https://www.w3.org/TR/SVG/implnote.html#ArcImplementationNotes)
in the subsection [F.6.2 Out-of-range
parameters](https://www.w3.org/TR/SVG/implnote.html#ArcOutOfRangeParameters).

The rx/ry arguments grammar is changed to `number` from
`nonnegative_number`
and then the rule's inline code converts to the absolute value, to
conform
to the implementation notes.

The two 'flag' values are changed to `number` from `flag`, and then the
rule's inline code forces nonzero values to 1.  As this was the only use
of the rule `flag` that rule is deleted from the grammar.